ready-to-use arena script

+++ arena.sh
@@ -2,21 +2,58 @@
 
 set -ue
 
-field=$1
 
+die() {
+  echo $1
+  exit 1
+}
+
+abspath() {
+  D=`dirname "$1"`
+  B=`basename "$1"`
+  echo "`cd \"$D\" 2>/dev/null && pwd || echo \"$D\"`/$B"
+}
+
+first=$(abspath "$1")
+second=$(abspath "$2")
+field=$3
+
+INFO=../info
+
+[ ! -x $first  ] && die "Bad first player"
+[ ! -x $second ] && die "Bad second player"
+[ ! -f $field  ] && die "Bad field"
+
+echo -e "First player:\t$first"
+echo -e "Second player:\t$second"
+
+
+echo "Preparing arena..."
 mkdir -p arena
 rm -f arena/*
 cp $field arena/matrix.txt
 
 cd arena
-
 echo "Start:" 
-cat matrix.txt 
-echo
 
-while [[ $(head -n 1 matrix.txt | cut -d\  -f3) == 'U' ]]; do
-  ../bot
-  cat matrix.txt
+while true; do
+  eval $($INFO --info matrix.txt)
+
+  $INFO --pretty matrix.txt
   echo
-  sleep 1
+
+  [ x$state != xU ] && break
+  [ $turn -gt 60  ] && break
+
+  rem=$((turn%2))
+  if [ $rem -eq 1 ]; then
+    this_player=A
+    this_player_exe="$first"
+  else
+    this_player=B
+    this_player_exe="$second"
+  fi
+
+  echo "Turn $((1 + turn/2))/$this_player:"
+  $this_player_exe
 done
